@@ -8,7 +8,7 @@
 
     <groupId>com.sap.cloud.s4-sample-ext</groupId>
     <artifactId>extorderapp</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>0.0.1-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <dependencyManagement>
@@ -16,7 +16,7 @@
             <dependency>
                 <groupId>com.sap.cloud.s4hana</groupId>
                 <artifactId>sdk-bom</artifactId>
-                <version>1.10.0</version>
+                <version>2.20.1</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
@@ -24,7 +24,7 @@
     </dependencyManagement>
 
     <properties>
-        <java.version>1.7</java.version>
+        <java.version>1.8</java.version>
 
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
@@ -47,17 +47,17 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-enforcer-plugin</artifactId>
-                <version>3.0.0-M1</version>
+                <version>3.0.0-M2</version>
                 <executions>
                     <execution>
-                        <id>S/4HANA Cloud SDK Project Structure Checks</id>
+                        <id>SAP Cloud SDK Project Structure Checks</id>
                         <goals>
                             <goal>enforce</goal>
                         </goals>
                         <configuration>
                             <rules>
                                 <requireMavenVersion>
-                                    <version>3.3.9</version>
+                                    <version>3.5</version>
                                 </requireMavenVersion>
                                 <requireJavaVersion>
                                     <version>${java.version}</version>
